Add some constants to types
