+++ sacloud/types/archive_size.go
@@ -1,0 +1,4 @@
+package types
+
+// ValidArchiveSizes 有効なアーカイブのサイズ
+var ValidArchiveSizes = []int{20, 40, 60, 80, 100, 250, 500, 750, 1024}
